Ammo count works and empty click
The ammo count in the UI now properly changes when a shell is fired. Could make it in future that the alt fire doesn't have a seperate ammo type and just uses twice the number of shells. Made it so that when you are out of shells the gun clicks to show that it is empty. Cried quite a lot and this shouldn't have taken me as long as it did.

:(