Shop, coins, upgrade
Created coins that get retrieved when the player kills an enemy. Can be made to get coins when player finishes level too. Created a shop widget. Created a fully working upgrade that allows the player to upgrade the shotgun power. Currently debugs are plugged in as will be doing more work on this after lunch. Created a save of data across levels.